Create new file for decorat class
